major overhaul + all bug fixes
#2 fixed
Changed how we handle the constrained behavior globally. Thanks, @ZenithVal, for the idea.
We now lock the y-axis on the container and close all on the world constraint. This resolves #2 and makes other fixes easier.
Also, thanks for saying use two-frame animation. It seems to make things slightly more consistent.

This update also includes a debug panel, but not in the package, as it is unnecessary. I used it to debug how remote client behavior is janky to visualize my animator from the remote client in VRChat.
The timings are reworked again and now we check for y velocity before resetting the floor collider.
This honestly looks fantistic on desktop somehow and has little jittering at all now. This was not tested in vr as its very late :)
Honestly there likely wont be another major change unless this can be done more efficiently. Minor tweaks to timings at best or better bhop detection maybe next.